payments and received_payments are separated now
